requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-groq
langchain-text-splitters>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
langchain-chroma
langchain-huggingface
PyGithub==2.2.0
python-dotenv==1.0.1
pydantic>=2.0.0
tiktoken
sentence-transformers
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0